+++ polls/tests/test_index.py
@@ -1,0 +1,61 @@
+import datetime
+from django.test import TestCase
+from django.urls import reverse
+from django.utils import timezone
+from polls.models import Question
+
+
+def create_question(question_text, days):
+    """Return Question."""
+    time = timezone.now() + datetime.timedelta(days=days)
+    time1 = timezone.now() + datetime.timedelta(days=days+7)
+    return Question.objects.create(question_text=question_text,
+                                   pub_date=time,
+                                   end_date=time1)
+
+
+class QuestionIndexViewTests(TestCase):
+    """Class to test all functions in index."""
+
+    def test_no_questions(self):
+        """Test index is no question."""
+        response = self.client.get(reverse('polls:index'))
+        self.assertEqual(response.status_code, 200)
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_past_question(self):
+        """Test have a question in past."""
+        create_question(question_text="Past question.", days=-30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_future_question(self):
+        """Test have question published in future."""
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertContains(response, "No polls are available.")
+        self.assertQuerysetEqual(response.context['latest_question_list'], [])
+
+    def test_future_question_and_past_question(self):
+        """Test have both past and future questions."""
+        create_question(question_text="Past question.", days=-30)
+        create_question(question_text="Future question.", days=30)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question.>']
+        )
+
+    def test_two_past_questions(self):
+        """Test have more than one questions in past."""
+        create_question(question_text="Past question 1.", days=-30)
+        create_question(question_text="Past question 2.", days=-5)
+        response = self.client.get(reverse('polls:index'))
+        self.assertQuerysetEqual(
+            response.context['latest_question_list'],
+            ['<Question: Past question 2.>', '<Question: Past question 1.>']
+        )
